Merge in SECSWST/spyglass from develop to develop-ota

* commit '251e00e2eb0892fe4910518d740336870a223562': (23 commits)
  HCS: Removed unused member from DispatcherMessage
  HCS: StrataPlatformMessage renamed to DispatcherMessage
  CS-1799: Platform - PlatformMessageData.h moved to src folder
  CS-1799: Removed research spike for proving of PlatformMessage concept
  CS-1799: Added comments into PlatformMessage header file
  Renamed 'PlatformMessage' to 'StrataPlatformMessage' in dispatcher
  CS-1799: Integrated PlatformMessage into Platform signals
  windows: set libzmq binary path to be on external path
  MacOSX: Set libzmq binary directory to be on external path
  cmake: hide duplicate missing tag complains on project configuration
  Add copy step for Mac OSX logic
  improve libzmq  dynamic library copying to bin by addign custom external project step
  Fixed Graying out issue
  Login: prevent window close in some cases to allow extra time for session close request
  Remove all refs to old CouchbaseLiteCpp
  added condition to check if APPS_CORESW_SDS_PLUGIN_CVC is disabled
  Remove couchbase from SCI
  update wording of the comment
  changed qDebug message
  fixed indentation
  ...

+++ libs/Board/Platform/src/PlatformMessageData.cpp
@@ -19,7 +19,7 @@ PlatformMessageData::PlatformMessageData(const QByteArray& rawMessage)
 
     if (result.IsError()) {
         jsonErrorString_ = rapidjson::GetParseError_En(result.Code());
-        jsonErrorOffset_ = static_cast<uint>(result.Offset());
+        jsonErrorOffset_ = static_cast<ulong>(result.Offset());
 
         qCWarning(logCategoryPlatformMessage).nospace().noquote()
             << QStringLiteral("JSON parse error at offset ") << jsonErrorOffset_
